Add ability to install packages after OS install
Support package managers being used to download packages, their
dependencies and install them on the target after the OS has been
installed.

Testing requires a repo file for the package manager being used
(currently only zypper is supported) and a repository located in
repos/$packagemanager/ that has the linux package (kernel) in it.