The has_answer attribute of the Confirm widget now used consistently
For the Confirm widget, the has_answer attribute was set to
True or False depending on what was clicked. However, the
actual label of the button pressed is stored in the 'response'
attribute. It would seem the 'has_answer' attribute should indicate
the existance of a response, yet in some places the has_answer
attribute was being used as the actual answer. So when I made an
earlier change for has_answer to be set to True in all cases
suddenly the "no" option for a bunch of dialog boxes unexpectedly
broke. All Confirm dialog-box choice examination has been changed
to inspect the 'response' attribute.