Update ister for swupd config file changes
swupd now uses a configuration file that contains version url, content
url and format id. Change ister to default to the stable url and read
the format file from disk (if not given on the commandline).

Probably need to request swupd to add a command to output its
format/version url/content url to avoid reading swupd configuration
files which may change in the future though.