Simplify keypress navigation
Remove the ability to navigate 'up' or 'shift tab' between widgets.
These keys are handled by urwid in a way that causes our focus to lose
track of its location. This causes the focus to jump in unexpected and
unpredictable ways. 'shift tab' and 'up' still work within widgets, such
as the FormBody and NavBar widgets. 'tab' and 'right arrow' still work
in the expected way, allowing the user to navigate between widgets.

In the NavBar, the 'Next' button is now the first button to receive
focus, optimizing the "configure -> next -> configure -> next" workflow.

Signed-off-by: Matthew Johnson <matthew.johnson@intel.com>