Add --dry-run argument option
When ister_gui.py exits, it calls either `reboot` or `poweroff`, even
when it is run locally with `sudo python ister_gui.py`. Unless
interrupted with CTRL-C, the installer shuts down the host on exit.
This adds a command line argument (-d, --dry-run) that specifies
that no install should run and the system should not reboot or
poweroff on exit.

Signed-off-by: Matthew Johnson <matthew.johnson@intel.com>