Add network check on splash screen
Display network requirements and give indication of whether
the system is able to connect to clearlinux.org or not. Allow the
user to configure a static ip address, gateway, and interface for the
installer to use.

Also add configurable proxy to installer gui and pass to swupd via the
json template. This way, if pacrunner is not working the user can
define their own proxies to connect to clearlinux.org for the
installation.

Signed-off-by: Matthew Johnson <matthew.johnson@intel.com>