Add checks for BundleSelectorStep arguments
This class can be instantiated for the sole purpose of retreiving a
list of required bundles. In this case it should not be necessary to
pass cur_step and tot_steps arguments since they will not be used.

This sets default values of None and only builds the progress widget
if values have been passed to the constructor. This prevents a crash
when the installation confirmation step is reached.

Signed-off-by: Matthew Johnson <matthew.johnson@intel.com>