Show swupd --install output in GUI
Users often complain about ister_gui not giving any status indication
when swupd starts up during the last step of the installation. This
patch displays the swupd output in the GUI alert so users can at
least see that the process is not frozen.

This patch also updates run_command to poll the running process instead
of waiting on it, which could cause deadlock if process output is too
long:

https://docs.python.org/2/library/subprocess.html#subprocess.Popen.wait

Signed-off-by: Matthew Johnson <matthew.johnson@intel.com>