Add first screen introduction to Clear Installer
After UX review, users were confused and intimidated by the first
screen being a network configuration screen. This adds a screen that
informs the user they are installing Clear Linux and a brief message
about Clear.

Adding this screen required an update to the progress messages at the
top of each consecutive screen.

Signed-off-by: Matthew Johnson <matthew.johnson@intel.com>