Allow user to continue with undetected interface settings
This patch allows the user to continue to the next screen if they
modified the interface/ip/gateway information unintentially and
did not attempt to set it.

The interface settings should not restrict the user from advancing
to the next screen in the installer. Even if there is no network
connection, they could still choose an action other than install
on the ChooseAction screen.

Signed-off-by: Matthew Johnson <matthew.johnson@intel.com>