Redesign network screen to group like config options
This was a confusing screen because it was not clear which buttons
were related to which configuration field. They are not grouped in
a logical manner and the < reset network > button only appears when
it is relevant.

Also remove the .8 second sleep on startup since this is no longer
the first screen. The time the user spends on the splash screen will
allow the network units to start up.

Signed-off-by: Matthew Johnson <matthew.johnson@intel.com>