Dynamically set button padding on message length
All regular screen, navbar, and alert buttons are padded based to the
length of the message they contain. This is achieved with a wrapper
function for the three methods that are always used to create ister
buttons (urwid.Button, urwid.AttrMap, and urwid.Padding). This
function, ister_button, also handles assigning click signals to the
buttons. The result is a consistent button style looking like this,

< button >

instead of this,

< button                                                             >

Signed-off-by: Matthew Johnson <matthew.johnson@intel.com>