Move static network validation to user edits
As the user edits the static IP configuration boxes, their input is
checked in real-time as they type. Once the field they are editing is
detected as a valid interface/IP address, the field caption turns
green, notifying the user of a valid configuration. The button to set
the static configuration does not activate until all required fields
are valid.

Moving this logic to before button activation allows the removal of a
lot of error handling complexity in the handler() function since the
user is not allowed to press the button until the input is valid.
Because of this, this patch also includes considerable simplification
to the NetworkRequirements class.

Signed-off-by: Matthew Johnson <matthew.johnson@intel.com>