Enable expect for automatic and manual installation
Includes an update of original autoinstall.expect commands for new gui
and updates to validate_release.py and helper scripts/config files to
pull the latest version of Clear Linux.

The new expect script runs through the manual installation steps in the
ister gui. The script is run via the same validate_release.py script as
the automatic installation script.

Signed-off-by: Matthew Johnson <matthew.johnson@intel.com>