Reduce required bundles and add default bundles
Remove network-basic and sysadmin-basic bundles from the required
bundles list and add them to a "default bundles" list. The default
bundles are selected by default but the user is able to deselect them.

This keeps the "required" size of the OS small, but indicates to the
user which bundles are commonly used because they are selected by
default.

Signed-off-by: Matthew Johnson <matthew.johnson@intel.com>