Make use of swupd's --path option for repair
The 'Repair' option in Ister was failing to complete the repair due to a
failed network connection in the chroot. Instead of chrooting and trying
to set up a network, the path to the mounted host OS can be provided to
swupd via the --path option. This removes the complicated chroot
requirements and fixes the 'Repair' option.

Signed-off-by: Matthew Johnson <matthew.johnson@intel.com>