Require sysadmin-basic bundle when admin user created
When an administrated user is created, root login is disabled. For this
user to be able to do much of anything in the system, they need sudo.
sudo will not be included unless the sysadmin-basic bundle is selected.
This patch makes sysadmin-basic required only when an administrative
user is set, and makes it optional when a user is not set (login as
root).

Ister also remembers the user's last selection before the bundle was
made mandatory. This allows the user to change their mind about user
configuration without worrying about deselecting the sysadmin-basic
bundle again.

Because the required_bundles list must be generated from user
configuration, it is separated out into its own function. This function
can then be called from the final step before the automatic install that
adds required bundles to the bundle list.

Signed-off-by: Matthew Johnson <matthew.johnson@intel.com>