Add sudo users to the sudo group, not to the sudoers config file
The general expectation for sudo enabling is that authentication is
required by way of being in the sudo (wheel) group. Consequently this
also serves to help in terms of security by requiring a password, and
to prevent accidental dangerous commands (`sudo rm -rf`, for example.)

With this change, we now place the newly created user into the wheel
group, and no longer create the sudoers.d/* files.

Signed-off-by: Ikey Doherty <michael.i.doherty@intel.com>