Wait for network service on first network screen load
The first time a user reaches the NetworkRequirements screen, poll
systemctl status systemd-networkd for an "Active" message.

This fixes the issue where a user tabs past the first screen fast
enough that the network service had not loaded yet, causing the screen
to display a "not connected" message. Now the first time a user reaches
that screen the network service will be started and any "not connected"
message will be a result of an actual network test.

Signed-off-by: Matthew Johnson <matthew.johnson@intel.com>