Rework statedir handling for physical installs
When doing physical installations, the statedir argument was being
largely ignored incorrectly due to the use of the bind mount. Rework the
bind mount usage and at the same time fixup the creation of the
directory to set the mode to only be accessable by the owner (root).
Also fixup cleanup to handle the potential new location of the statedir.