Allow configuration to dictate network testing
Instead of always testing the clearlinux website as the network
available standard, make use of the content and version urls passed in
as arguments. At the same time update the clearlinux urls to reflect the
cdn to the update content instead of the main page similar to how the
swupd-client itself works.