Add interface configuration field for target
Add configuration field for the interface pattern in target static
network configuration screen. This allows users to specify which NIC the
static configuration is for on machines with multiple NICs. Without this
feature users were seeing the same static IP address being set for every
NIC on the server, which caused connection failures.

Signed-off-by: Matthew Johnson <matthew.johnson@intel.com>