Read available disks every time screen is reached
Instead of reading the list of disks only once when ister_gui is
initializing, read them every time the disk partition screen is reached.
This fixes a bug where slow-loading devices have not loaded by the time
the available disks are determined and are therefore never discovered by
ister, causing the installation to be impossible.

The choice was made to read the disks every time instead of just until
some devices are found since some devices may load faster than others.

Also add a "Refresh" button to allow users to refresh the screen in the
event they add a device.

Signed-off-by: Matthew Johnson <matthew.johnson@intel.com>