Handle manual swap partition detection
In the gui swap manual partitions created with cgdisk were not getting
detected and added to the template (causing ister to not format the
partitions and then systemd to fail to activate the swap
partition). To resolve this rework the detection process to look at
all the items chosen in cgdisk and detect for only those without a
mount point if their filesystem type code was set to be a "Linux swap"
partition.