Add option for moving statedir into targetdir
The current install approach favors disk space by allowing for the state
dir to exist outside of the target dir.  However, the two directories
can exist across block device boundaries.  To favor time to install,
adding an option to relocate the state dir into the target dir so they
are located within the same block device.  This can reduce install time
by as much as 60%.  Once the install has completed, the state dir is
removed from the target dir.

This has a couple of side effects.  When this option is used, the
template needs to allocate space for the state dir in addition to actual
content.  Additionally, this option cannot be used in conjunction with
the statedir option, as the state dir in this case is automatically
derived.  Because of these side effects, this option is not enabled by
default.

Signed-off-by: George T Kramer <george.t.kramer@intel.com>