Add support for installing content with dnf
In test scenarios, work to create a swupd update stream is not needed
and increases image generation time.  If content for a test scenario can
be supplied as a package repository, then dnf can be used to do an
install of packages in the same way mixer generates chroots.

To support a define-and-run approach with ister, template files may
define an additional "SoftwareManager" key, which may be either "dnf" or
"swupd", if not defined to preserve backwards compatbility.  In the case
where the package manager is swupd, the list of bundles remains the same
as it is today.  In the case where the package manager is dnf, the
bundle list is treated as a package list that is used during a dnf
install.

Much like how content and version URLs may be passed at the command line
for swupd, to pass package repository URLs, which there may be multiple,
a DNF configuration file must be passed instead with the new
--dnf-config option.

To mimic how swupd works after installing content, update triggers are
run if they exist.  Because using DNF makes ister distro-agnostic,
opting to make the image bootable by using a post non-chroot script as
opposed to calling clr-boot-manager update directly.  The
dnf-post-update.sh script has been included to generate a bootable image
after usage of DNF.

Signed-off-by: George T Kramer <george.t.kramer@intel.com>