ister.py: enhance the isterconf= kernel cmdline option
The 'isterconf=' kernel command-line option is interpreted by ister as a link
to the ister configuration file. Ister downloads it, reads and expects to see

template=<URL>

at the first line. And the URL is the ister.json template file URL. Ister
downloads it and installs ClearLinux accordingly.

Now, my task is to pass ister the 'ister.json' URL via command line. To do
that, I need to have 2 files - the configuration file and the 'ister.json'
file. This is overly complex for my task and this patch offers an improvement.

Lets teach ister to be flexible and allow passing both the configuration file
and the template file using the 'isterconf' option. Ister can just read
whatever it is given, check if it looks like a configuration file, and if it is
not, treat it as the template file (ister.json) instead of just erroring out.
This is exactly what the patch does.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>