ister.py: generalize FS utility map
This patch is a preparation for the next patch where we'll add the label
support to ister. Before doing that we need to make the 'fs_util' dictionary a
bit more general and not only mape FS type to FS util name, but map FS type to
multiple things.

So this patch turns the plain 'fs_util' dictionay to a dictionary of
dictionaries so that we could store more FS-specific things besides the format
tool name. E.g., we can store the label option name there too.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>