ister_gui: Fix bug during Bundles initialization
The list is populate during the creation of the installation object with
the bundle list provided by the template found at /etc/ister.json,
inject there during the manufacturing of the installer image. Just after
that is was being cleared and therefore the templates list was never
respected.

That initialization is enough for the code to properly work since at
that point the code is executed only once, no matter if the user rolled
back some screens.

Signed-off-by: Murilo Belluzzo <murilo.belluzzo@intel.com>