ister: pylint cleanups part 1
Pylint is a useful tool for verifying python code. It often spots real errors.
Pylint complaints have different classes, like R for recommendation and C for
coding syle, E for errors, etc.

Some of the C and R complaints are often annoying and there is a way to shut
them down.

When I contribute to ister I ususally verify my changes with pylint, but this
is very difficult to do because there so many complaints for the current code
that the complaints my code introduces is hard to notice.

This patch addresses all the pylint complaints of the "wrong continued
indentetion" type, for example:

C:1603, 0: Wrong continued indentation (remove 1 space).
                        help="Path to swupd state dir")
                       |^ (bad-continuation)

I introduced temporary variables in some places to make it easier to fit into
80 lines.

Other types of pylint complaints will be addressed sepaately.

P.S. Generally python programs use 100 characters per line, not 80. But ister
uses 80 and it makes it more difficult to deal with string formatting.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>