ister: do not fail if user already exists
Ister allows specifying users in 'ister.json' - they will be created after
'swupd' installs all the bundles. E.g., one may include something like this to
'ister.json"

'Users' : [{'username': 'john', 'uid' : '1005', 'password' : '<hash>', 'key' : '<pubkey>'}]

and user 'john' will be created and and configured.

However, if user 'john' already exists, ister currently fails instead of just
continuing to configure UID, password, and the SSH public key for this user.
The reason the user may exist is it has been added at the bundle installation
phase. One example of such a user is 'root'.

This patch makes 'ister' to not fail if the user already exists. One practical
example is configuring root SSH key:

'Users' : [{'username': 'root', 'password' : '<hash>', 'key' : '<pubkey>'}]

In my case I have a simple SUT (system under test) in an isolated network, and
the strict ClearLinux security defaults are not suitable for my case. With this
patch I can instruct ister to configure a pre-defined root password and my root
ssh key, and it works just fine.

This patch uses the fact that 'useradd' and 'usermod' have very similar CLI and
option names are very consistent. So we go ahead and try to create the user
with 'useradd', and if it fails with the documented error code 9, we use
'usermod' tool instead. This tool will just modify the password and UID of the
existing user without creating this user.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>