ister_test: make the run_command wrapper return something
The 'run_command()' function in ister.py returns the (stdout, stderr, exitcode)
tuple, where 'stdout' and 'stderr' are per-line lists.

However, the 'run_command_wrapper()' in the 'ister_test.py' unit test program
is not consistent with that and does not return anything, which defaults to
returning 'None'.

Let's teach it to return someting a bit more consistent with 'run_command()'.
Let's at least return a tuple of two elements. I chose ([], [], 0): no output
at all and success.

This patch is necessary for the next patch where we'll teach ister to not fail
when creating a user that already exists.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>