Simplify check_kernel_cmdline
It is a bit silly to return a tuple like (True, path) or (False, ''). It is
enough to just return a path in case the ister.conf was found and 'None'
otherwise. This is a pretty standard practice.

Simplify the code and adjust the testcases.

On top of this, improve the 'check_kernel_cmdline()' commentery.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>