Smarter ister.conf download
It is inoptimal to have an unconditional 15 seconds delay before downloading
the ister.conf file. This patch improves it and turns the unconditional sleep
into a timeout - we keep trying for 15 seconds. If we succeed faster, we
continue. So the end result is that we generally install the OS faster. Hence,
this is an improvement.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>