ister_test: improve tracability
If a test case fails, we are often left with a vague question like "command xyz
filed", and it is sometime very hard to understand why it failed because
command xyz does tons of different things. This patch aims to improve the
situation and teaches the test runner to print the stack trace in case of a
failure. This way debugging becomes a lot easier process because the stack
trace points you exactly the right place. I do not see a reason to hide it and
make our lives more difficult.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>