Enhance run_command()
'run_command()' is one of the central functions in ister, used all over the
place. It runs an external command, wait for it to finish, and
prints/logs/returns its output.

Unfortunately 'run_command()' has a serious limitation - it always first waits
for all the stdout of the command, and then waits for stderr of the command.

And there are 2 problems with this.
1. A possibility of deadlock, documented in the code.
2. We do not see stderr until we receive all stdout.

The latter is very visible with 'swupd'. It prints stuff like
"Downloading packs" to stderr, but when we follow ister log - we do not see
this output because it is all blocked by us waiting for stdout to exhaust.

At the very end, we see all the stdout (many dots in case of swupd) and then
all the stderr. No the same as one would see if there was a simple file
redirection.

Well, and we are also lucky stderr output of swupd is not that large. If it was
larger and exceeded the pipe size, we'd deadlock.

This patch fixes the issue. It adds quite a bit of code, but the logic is this.

We create 2 threads, one is fetching stdout and the other is fetching stderr.
In parallel. The output is put to a queue, and the main process just reads the
queue and prints the output.

This fixes both of the above problems.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>