Do not create empty password by default
If ister.json declares a user that has to be created, but does not specify the
passord, ister creates it with an empty string password. This is not very secure
because users may just forget adding the password and assume ister behaves
similarly to the common tools like 'useradd': they create a user that cannot
login to the system because it does not have any passowrd set. In case of
ister, the user will have a password, which happens to be and empty string, and
this user can login.

This patch addresses the problem.

Also ajust the existing test cases and add a couple of new testcates to verify
creation of users with empty password and non-empty passwords.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>