ister: fix bug in existing user creation
If 'ister.json' has the instructions to create a user, but it already exists,
we fall back to using 'usermod' instead of 'useradd'. However, if no addinional
options like 'uid' or 'password' were given, we end up running:

usermod <username>

without any options like -u or -p, and this fails, because we did not specify
what shold be modified.

This patch fixes the issue by checking whether there are some additional
options to use. With this patch we will not run 'usermod <username>', but we
will run something like "usermod -u <uid> <username>".

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>