ister: improve standard mount paths detection
Ister tries to avoid generating mount units for "standard" mount points,
assuming that systemd will do the right thing at first boot. One of the
"standard" mount points is "/usr" and any mountpoint inside "/usr".

The way ister was checking for this "standard" mountpoint had an issue
and it would treat something like "/usr-2" as a standard directory, which
is incorrect. This patch fixes the potential problem.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>